// src/constants/enums.ts
export enum NOTIFICATION_TYPE {
  ERROR = "error",
  INFO = "info",
  SUCCESS = "success",
  WARNING = "warning"
}

export enum COLORS {
  ERROR = "#F44336",
  INFO = "#3F51B5",
  SUCCESS = "#4CAF50",
  WARNING = "#FFC107",
  WHITE = "#FFFFFF"
}

export enum POSITION_FROM {
  TOP = "top",
  RIGHT = "right",
  BOTTOM = "bottom",
  LEFT = "left"
}